feat: split tests into multiple files
