+++ packages/core/src/controllers/relayer.ts
@@ -151,7 +151,6 @@ export class Relayer extends IRelayer {
   public async transportClose() {
     this.transportExplicitlyClosed = true;
     if (this.connected) await this.provider.disconnect();
-    await new Promise<void>((resolve) => setTimeout(resolve, 2000));
     this.events.emit(RELAYER_EVENTS.transport_closed);
     console.log("transport closed --- @!", this.core.name);
   }
@@ -194,6 +193,7 @@ export class Relayer extends IRelayer {
 
   public async restartTransport(relayUrl?: string) {
     await this.transportClose();
+    await new Promise<void>((resolve) => setTimeout(resolve, 2000));
     await this.transportOpen(relayUrl).catch((e) => {
       console.log("restartTransport error", e);
     });
