+++ packages/sign-client/test/sdk/client.spec.ts
@@ -1,5 +1,5 @@
-import { getSdkError, generateRandomBytes32 } from "@walletconnect/utils";
-import { expect, describe, it, vi, beforeAll, afterEach } from "vitest";
+import { getSdkError } from "@walletconnect/utils";
+import { expect, describe, it, vi } from "vitest";
 import SignClient from "../../src";
 import {
   initTwoClients,
@@ -8,9 +8,6 @@ import {
   deleteClients,
   throttle,
 } from "../shared";
-
-const generateClientDbName = (prefix: string) =>
-  `./test/tmp/${prefix}_${generateRandomBytes32()}.db`;
 
 describe("Sign Client Integration", () => {
   it("init", async () => {
@@ -128,69 +125,6 @@ describe("Sign Client Integration", () => {
           await deleteClients(clients);
         });
       });
-      describe("after restart", () => {
-        it("clients can ping each other", async () => {
-          const db_a = generateClientDbName("client_a");
-          const db_b = generateClientDbName("client_b");
-
-          let clients = await initTwoClients(
-            {
-              name: "pairing -> after restart A before",
-              storageOptions: { database: db_a },
-            },
-            {
-              name: "pairing -> after restart B before",
-              storageOptions: { database: db_b },
-            },
-            { logger: "error" },
-          );
-          const {
-            pairingA: { topic },
-          } = await testConnectMethod(clients);
-
-          await Promise.all([
-            new Promise((resolve) => {
-              // ping
-              clients.B.core.pairing.events.on("pairing_ping", (event: any) => {
-                resolve(event);
-              });
-            }),
-            new Promise((resolve) => {
-              clients.A.core.pairing.events.on("pairing_ping", (event: any) => {
-                resolve(event);
-              });
-            }),
-            new Promise(async (resolve) => {
-              // ping
-              await clients.A.ping({ topic });
-              await clients.B.ping({ topic });
-              resolve(true);
-            }),
-          ]);
-
-          await deleteClients(clients);
-
-          await throttle(1_000);
-          // restart
-          clients = await initTwoClients(
-            {
-              name: "pairing -> after restart A after",
-              storageOptions: { database: db_a },
-            },
-            {
-              name: "pairing -> after restart B after",
-              storageOptions: { database: db_b },
-            },
-            { logger: "error" },
-          );
-
-          // ping
-          await clients.A.ping({ topic });
-          await clients.B.ping({ topic });
-
-          await deleteClients(clients);
-        });
-      });
     });
     describe("session", () => {
       describe("with existing session", () => {
@@ -214,67 +148,6 @@ describe("Sign Client Integration", () => {
             sessionA: { topic },
           } = await testConnectMethod(clients);
           await clients.B.ping({ topic });
-          await deleteClients(clients);
-        });
-      });
-      describe("after restart", () => {
-        it("clients can ping each other", async () => {
-          const db_a = generateClientDbName("client_a");
-          const db_b = generateClientDbName("client_b");
-          let clients = await initTwoClients(
-            {
-              name: "session -> after restart A before",
-              storageOptions: { database: db_a },
-            },
-            {
-              name: "session -> after restart B before",
-              storageOptions: { database: db_b },
-            },
-          );
-          const {
-            sessionA: { topic },
-          } = await testConnectMethod(clients);
-
-          await Promise.all([
-            new Promise((resolve) => {
-              // ping
-              clients.B.on("session_ping", (event: any) => {
-                resolve(event);
-              });
-            }),
-            new Promise((resolve) => {
-              clients.A.on("session_ping", (event: any) => {
-                resolve(event);
-              });
-            }),
-            new Promise(async (resolve) => {
-              // ping
-              await clients.A.ping({ topic });
-              await clients.B.ping({ topic });
-              resolve(true);
-            }),
-          ]);
-
-          // delete
-          await deleteClients(clients);
-
-          await throttle(1_000);
-          // restart
-          clients = await initTwoClients(
-            {
-              name: "session -> after restart A after",
-              storageOptions: { database: db_a },
-            },
-            {
-              name: "session -> after restart B after",
-              storageOptions: { database: db_b },
-            },
-          );
-
-          // ping
-          await clients.A.ping({ topic });
-          await clients.B.ping({ topic });
-          // delete
           await deleteClients(clients);
         });
       });
@@ -307,7 +180,7 @@ describe("Sign Client Integration", () => {
       const result = clients.A.session.get(topic).namespaces;
       expect(result).to.eql(namespacesAfter);
       await deleteClients(clients);
-    }, 50_000);
+    });
   });
 
   describe("extend", () => {
@@ -331,80 +204,6 @@ describe("Sign Client Integration", () => {
       const updatedExpiry = clients.A.session.get(topic).expiry;
       expect(updatedExpiry).to.be.greaterThan(prevExpiry);
       vi.useRealTimers();
-      await deleteClients(clients);
-    }, 50_000);
-  });
-  describe("transport", () => {
-    it("should disconnect & reestablish socket transport", async () => {
-      const clients = await initTwoClients(
-        { name: "disconnect & reestablish socket transport A" },
-        { name: "disconnect & reestablish socket transpor B" },
-      );
-      const {
-        sessionA: { topic },
-      } = await testConnectMethod(clients);
-
-      console.log("closing transports");
-      await clients.A.core.relayer.restartTransport();
-      await throttle(500);
-      await clients.B.core.relayer.restartTransport();
-      console.log("opened transports");
-
-      await throttle(2000);
-      await Promise.all([
-        new Promise((resolve) => {
-          clients.B.on("session_ping", (event: any) => {
-            resolve(event);
-          });
-        }),
-        new Promise((resolve) => {
-          clients.A.on("session_ping", (event: any) => {
-            resolve(event);
-          });
-        }),
-        new Promise(async (resolve) => {
-          await clients.A.ping({ topic });
-          await clients.B.ping({ topic });
-          resolve(true);
-        }),
-      ]);
-      await deleteClients(clients);
-    });
-    it("should disconnect & reestablish socket transport with delay", async () => {
-      const clients = await initTwoClients(
-        { name: "disconnect & reestablish socket transport with delay A" },
-        { name: "disconnect & reestablish socket transport with delay B" },
-      );
-      const {
-        sessionA: { topic },
-      } = await testConnectMethod(clients);
-      console.log("closing transport A");
-      await clients.A.core.relayer.transportClose();
-      await throttle(2000);
-      console.log("opening transport A");
-      await clients.A.core.relayer.transportOpen();
-      console.log("closing transport B");
-      await clients.B.core.relayer.transportClose();
-      await throttle(2000);
-      console.log("opening transport B");
-      await clients.B.core.relayer.transportOpen();
-      await Promise.all([
-        new Promise((resolve) => {
-          clients.B.on("session_ping", (event: any) => {
-            resolve(event);
-          });
-        }),
-        new Promise((resolve) => {
-          clients.A.on("session_ping", (event: any) => {
-            resolve(event);
-          });
-        }),
-        new Promise(async (resolve) => {
-          await clients.A.ping({ topic });
-          await clients.B.ping({ topic });
-          resolve(true);
-        }),
-      ]);
       await deleteClients(clients);
     });
   });
